Move InboxSize to heph::future
To be added as ActorFuture option.
