+++ rt/src/spawn/options.rs
@@ -6,11 +6,10 @@
 //! [`Future`]: std::future::Future
 
 use std::cmp::Ordering;
-use std::error::Error;
-use std::fmt;
-use std::num::NonZeroU8;
 use std::ops::Mul;
 use std::time::Duration;
+
+pub use heph::future::InboxSize;
 
 /// Options for [spawning] an [`Actor`].
 ///
@@ -47,7 +46,7 @@ impl ActorOptions {
     /// Default options for system actors.
     pub(crate) const SYSTEM: ActorOptions = ActorOptions {
         priority: Priority::SYSTEM,
-        inbox_size: InboxSize::SYSTEM,
+        inbox_size: InboxSize::ONE,
     };
 
     /// Returns the priority set in the options.
@@ -164,69 +163,6 @@ fn priority_duration_multiplication() {
     assert!(high < low);
     assert!(normal < low);
 }
-
-/// Maximum size of the actor's inbox before adding into it will
-/// (asynchronously) block the sender.
-///
-/// # Notes
-///
-/// The sizes defines by the constants may change in the future.
-#[derive(Copy, Clone, Debug)]
-pub struct InboxSize(NonZeroU8);
-
-impl InboxSize {
-    /// Space for a single message.
-    pub const ONE: InboxSize = InboxSize(NonZeroU8::new(1).unwrap());
-
-    /// Small inbox, currently 8 messages.
-    pub const SMALL: InboxSize = InboxSize(NonZeroU8::new(8).unwrap());
-
-    /// Medium sized inbox, currently 16 messages.
-    pub const MEDIUM: InboxSize = InboxSize(NonZeroU8::new(16).unwrap());
-
-    /// Maximum inbox size, currently 24 messages.
-    pub const LARGE: InboxSize = InboxSize(NonZeroU8::new(24).unwrap());
-
-    /// Maximum inbox size, currently 29 messages.
-    pub const MAX: InboxSize = InboxSize(NonZeroU8::new(heph_inbox::MAX_CAP as u8).unwrap());
-
-    /// System priority.
-    const SYSTEM: InboxSize = InboxSize::ONE;
-
-    /// Use a fixed inbox size.
-    ///
-    /// Returns an error if the inbox size is either zero or too big.
-    pub const fn fixed(size: usize) -> Result<InboxSize, InvalidInboxSize> {
-        if size >= 1 && size <= heph_inbox::MAX_CAP {
-            Ok(InboxSize(NonZeroU8::new(size as u8).unwrap()))
-        } else {
-            Err(InvalidInboxSize)
-        }
-    }
-}
-
-impl Default for InboxSize {
-    fn default() -> InboxSize {
-        InboxSize::SMALL
-    }
-}
-
-/// Error returned by [`InboxSize::fixed`].
-#[derive(Debug)]
-pub struct InvalidInboxSize;
-
-impl fmt::Display for InvalidInboxSize {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "actor inbox capacity must be between {} and {}",
-            heph_inbox::MIN_CAP,
-            heph_inbox::MAX_CAP
-        )
-    }
-}
-
-impl Error for InvalidInboxSize {}
 
 /// Options for spawning a [`SyncActor`].
 ///
